requirements.txt
# RAG Pipeline Requirements
google-genai>=1.0.0
pypdf2>=3.0.0
openpyxl>=3.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
